Quick fix

+++ 0x00-python_variable_annotations/0-add.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
 
 
-def add(a: float, b: int) -> float:
+def add(a: float, b: float) -> float:
     return float(sum([a, b]))
